Extract build_coverage_matrix into a shared file
Both appraisal/generate.rb and the new dependency:verify check need the
integration -> appraisal-name mapping build_coverage_matrix computes, so it
moves out of generate.rb into tasks/appraisal_coverage.rb where both can
require it without duplicating the logic.
